src/requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy==1.26.3
Pillow==10.4.0
matplotlib==3.7.2